move proc decorator to new file for use across all procs.
